replace non-standard uint with size_t

+++ examples/fcu_arm_test.cpp
@@ -22,7 +22,7 @@
 
 int main(int argc, char *argv[]) {
     const std::string device = (argc>1) ? std::string(argv[1]) : "/dev/ttyUSB0";
-    const uint baudrate = (argc>2) ? std::stoul(argv[2]) : 115200;
+    const size_t baudrate = (argc>2) ? std::stoul(argv[2]) : 115200;
 
     std::chrono::high_resolution_clock::time_point start, end;
     bool feature_changed = false;
