+++ inc/msp/MSP.hpp
@@ -12,7 +12,7 @@ namespace msp {
 
 // size of header+crc, e.g. amount of bytes in MSP message that do not belong to payload
 // preamble (2) + direction (1) + size (1) + command (1) + crc (1) = 6
-static const uint FRAME_SIZE = 6;
+static const size_t FRAME_SIZE = 6;
 
 // exception to throw when header contains wrong data
 class MalformedHeader : public std::runtime_error {
@@ -41,7 +41,7 @@ public:
     WrongCRC(const uint8_t msg_id, const uint8_t exp, const uint8_t rcv)
         : std::runtime_error(
               "CRC not matching: "
-              "Message "+std::to_string(uint(msg_id))+", "
+              "Message "+std::to_string(size_t(msg_id))+", "
               "expected CRC "+std::to_string(exp)+", "
               "received CRC "+std::to_string(rcv))
     { }
@@ -90,7 +90,7 @@ public:
      * @param device device path
      * @param baudrate serial baudrate
      */
-    MSP(const std::string &device, const uint baudrate=115200);
+    MSP(const std::string &device, const size_t baudrate=115200);
 
     /**
      * @brief connect establish connection to serial device
@@ -98,7 +98,7 @@ public:
      * @param baudrate serial baudrate (default: 115200)
      * @return true on success
      */
-    bool connect(const std::string &device, const uint baudrate=115200);
+    bool connect(const std::string &device, const size_t baudrate=115200);
 
     /**
      * @brief request send command and request data from FC once
@@ -122,7 +122,7 @@ public:
      * @param min_payload_size minimum amount of payload (bytes) that needs to be available before reading and decoding of a message starts
      * @return true when data has been received
      */
-    bool request_wait(msp::Request &request, const uint wait_ms, const uint min_payload_size = 0);
+    bool request_wait(msp::Request &request, const size_t wait_ms, const size_t min_payload_size = 0);
 
     /**
      * @brief respond send data to FC and read acknowledge
